[xharness] Fix 32 bit device tests selection
With the `run-ios-32-tests` label, all iOS device tests were ignored.
This is because the ignore condition was incorrect, `IncludeiOS` is not set for 32 bit device tests.
Now we only enable/disable the 32 bit device tests based on `IncludeiOS32` which is set by the `run-ios-32-tests` label.

IncludeiOS is now IncludeiOS64 and fix 'run-device-tests'

- Treat iOS 32 & 64 bit equally. This way you can use `run-ios-32-tests`, `run-ios-64-tests`, `skip-ios-32-tests` and `skip-ios-64-tests`.
- Also fix `run-device-tests`. I noticed that if I had: `skip-all-tests` and `run-device-tests` nothing happened (all tests ignored).
  This is because the skip all would make IncludeiOS64 false, leading to `Ignored = !IncludeiOS64` true.
  Since later on we did `task.Ignored |= ignored;`, even if `ignored` here, which is supposed to be `!IncludeDevice`, is false, we get `true || false == true` and we skip the device tests. (if you followed :P)
  Therefore I think we want `&=` to make the IncludeDevice work as expected.